requirements.txt
apiview==1.3.25
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.0.12
cloudinary==1.29.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.5
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-heroku==0.3.1
django-import-export==2.8.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc8
generic==1.1.0
gunicorn==20.1.0
idna==3.3
MarkupPy==1.14
Naked==0.1.31
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.1.1
psycopg2==2.9.3
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
routers==0.10.1
six==1.16.0
sqlparse==0.4.2
status==0.2.5
tablib==3.2.1
urllib3==1.26.9
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
